# Ugly datatable clicking functionality
This is a very ugly commit that enables you to click on datatables. They appear all over the place and the whole thing is really awful.

Also, not sure about where the comments appear. This is also pretty gross.

To do: finish this function, add DT button functionality to DTs, and sort out subdistricts and directorates to pull from database