# commit progress so far
I've added a few things to this test version. Interactive tables. Criticality.

I'm doing a root and branch code review. Every function is on the chopping block.

Likely to change every output and some of the interface.

This is the commit before I ruin everything, basically, but even this isn't really a working commit.